+++ Editor/src/com/kotcrab/vis/editor/ui/scene/entityproperties/autotable/StringStringMapView.java
@@ -26,7 +26,7 @@ import com.kotcrab.vis.editor.Icons;
 import com.kotcrab.vis.editor.ui.scene.entityproperties.EntityProperties;
 import com.kotcrab.vis.editor.util.scene2d.VisChangeListener;
 import com.kotcrab.vis.runtime.component.Variables;
-import com.kotcrab.vis.ui.InputValidator;
+import com.kotcrab.vis.ui.util.InputValidator;
 import com.kotcrab.vis.ui.util.dialog.DialogUtils;
 import com.kotcrab.vis.ui.widget.*;