+++ UI/src/com/kotcrab/vis/ui/util/Validators.java
@@ -16,7 +16,6 @@
 
 package com.kotcrab.vis.ui.util;
 
-import com.kotcrab.vis.ui.InputValidator;
 import com.kotcrab.vis.ui.util.dialog.DialogUtils;
 import com.kotcrab.vis.ui.widget.VisValidatableTextField;